nfts sort by name in CatalogCollection screen
